faser ut tp-offentlig steget og flytter sivilstand og samtykke steg
