--- src/components/stegvisning/OffentligTP/OffentligTP.module.scss
@@ -1,3 +1,0 @@
-.button {
-  margin: var(--a-spacing-6) var(--a-spacing-4) 0 0;
-}
