--- src/components/stegvisning/OffentligTP/OffentligTP.module.scss.d.ts
@@ -1,5 +1,0 @@
-import globalClassNames from "../../../style.d";
-declare const classNames: typeof globalClassNames & {
-  readonly button: "button";
-};
-export = classNames;
