--- src/components/stegvisning/OffentligTP/__tests__/OffentligTP.test.tsx
@@ -1,99 +1,0 @@
-import * as ReactRouterUtils from 'react-router'
-
-import { describe, it, vi } from 'vitest'
-
-import { OffentligTP } from '..'
-import { screen, render, waitFor, userEvent } from '@/test-utils'
-
-describe('stegvisning - OffentligTP', () => {
-  const onCancelMock = vi.fn()
-  const onPreviousMock = vi.fn()
-  const onNextMock = vi.fn()
-
-  it('rendrer slik den skal', async () => {
-    const result = render(
-      <OffentligTP
-        onCancel={onCancelMock}
-        onPrevious={onPreviousMock}
-        onNext={onNextMock}
-      />
-    )
-    expect(screen.getByRole('heading', { level: 2 })).toHaveTextContent(
-      'stegvisning.offentligtp.title'
-    )
-    await waitFor(() => {
-      expect(result.asFragment()).toMatchSnapshot()
-    })
-  })
-
-  it('kaller navigate når shouldRedirectTo er angitt', async () => {
-    const navigateMock = vi.fn()
-    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
-      () => navigateMock
-    )
-    const randomPath = '/random-path'
-
-    render(
-      <OffentligTP
-        shouldRedirectTo={randomPath}
-        onCancel={onCancelMock}
-        onPrevious={onPreviousMock}
-        onNext={onNextMock}
-      />
-    )
-    expect(navigateMock).toHaveBeenCalledWith(randomPath)
-  })
-
-  it('kaller onNext når brukeren klikker på Neste', async () => {
-    const user = userEvent.setup()
-    render(
-      <OffentligTP
-        onCancel={onCancelMock}
-        onPrevious={onPreviousMock}
-        onNext={onNextMock}
-      />
-    )
-    await user.click(screen.getByText('stegvisning.neste'))
-    expect(onNextMock).toHaveBeenCalled()
-  })
-
-  it('kaller onPrevious når brukeren klikker på Tilbake', async () => {
-    const user = userEvent.setup()
-    render(
-      <OffentligTP
-        onCancel={onCancelMock}
-        onPrevious={onPreviousMock}
-        onNext={onNextMock}
-      />
-    )
-
-    await user.click(screen.getByText('stegvisning.tilbake'))
-    expect(onPreviousMock).toHaveBeenCalled()
-  })
-
-  it('kaller onCancel når brukeren klikker på Avbryt', async () => {
-    const user = userEvent.setup()
-    render(
-      <OffentligTP
-        onCancel={onCancelMock}
-        onPrevious={onPreviousMock}
-        onNext={onNextMock}
-      />
-    )
-
-    await user.click(screen.getByText('stegvisning.avbryt'))
-    expect(screen.getByText('stegvisning.avbryt')).toBeInTheDocument()
-    expect(onCancelMock).toHaveBeenCalled()
-  })
-
-  it('viser ikke avbryt knapp når onCancel ikke er definert', async () => {
-    render(
-      <OffentligTP
-        onCancel={undefined}
-        onPrevious={onPreviousMock}
-        onNext={onNextMock}
-      />
-    )
-    expect(screen.queryByText('stegvisning.avbryt')).not.toBeInTheDocument()
-  })
-})
