+++ src/pages/Step3/__tests__/Step3.test.tsx
@@ -3,26 +3,20 @@ import { createMemoryRouter, RouterProvider } from 'react-router-dom'
 
 import { describe, it, vi } from 'vitest'
 
-import { fulfilledGetPerson } from '@/mocks/mockedRTKQueryApiCalls'
-import { mockErrorResponse } from '@/mocks/server'
+import {
+  fulfilledGetPerson,
+  rejectedGetInntekt,
+} from '@/mocks/mockedRTKQueryApiCalls'
 import { BASE_PATH, paths } from '@/router/constants'
 import { routes } from '@/router/routes'
 import { apiSlice } from '@/state/api/apiSlice'
 import { store } from '@/state/store'
 import * as userInputReducerUtils from '@/state/userInput/userInputReducer'
-import { screen, render, waitFor, userEvent } from '@/test-utils'
+import { screen, render, userEvent, waitFor } from '@/test-utils'
 
 const initialGetState = store.getState
 
 describe('Step 3', () => {
-  afterEach(() => {
-    store.dispatch(apiSlice.util.resetApiState())
-    vi.clearAllMocks()
-    vi.resetAllMocks()
-    vi.resetModules()
-    store.getState = initialGetState
-  })
-
   beforeEach(() => {
     store.getState = vi.fn().mockImplementation(() => ({
       api: {
@@ -32,167 +26,128 @@ describe('Step 3', () => {
       },
       userInput: {
         ...userInputReducerUtils.userInputInitialState,
-        samtykke: true,
       },
     }))
   })
 
-  describe('Gitt at kallet til /tpo-medlemskap er vellykket', async () => {
-    it('har riktig sidetittel', async () => {
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      await waitFor(async () => {
-        expect(document.title).toBe('application.title.stegvisning.step3')
-      })
+  afterEach(() => {
+    store.dispatch(apiSlice.util.resetApiState())
+    vi.clearAllMocks()
+    vi.resetAllMocks()
+    vi.resetModules()
+    store.getState = initialGetState
+  })
+
+  it('har riktig sidetittel og viser loader mens loaderen fetcher data', async () => {
+    store.getState = vi.fn().mockImplementation(() => ({
+      api: {
+        queries: {
+          ...rejectedGetInntekt,
+          ...fulfilledGetPerson,
+        },
+      },
+      userInput: {
+        ...userInputReducerUtils.userInputInitialState,
+      },
+    }))
+
+    const router = createMemoryRouter(routes, {
+      basename: BASE_PATH,
+      initialEntries: [`${BASE_PATH}${paths.afp}`],
+    })
+    render(<RouterProvider router={router} />, {
+      hasRouter: false,
     })
 
-    it('sender videre til steg 4 når brukeren klikker på Neste', async () => {
-      const user = userEvent.setup()
-      const navigateMock = vi.fn()
-      vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
-        () => navigateMock
-      )
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      expect(
-        await screen.findByText('stegvisning.offentligtp.title')
-      ).toBeVisible()
-
-      await user.click(await screen.findByText('stegvisning.neste'))
-
-      expect(navigateMock).toHaveBeenCalledWith(paths.afp)
-    })
-
-    it('sender tilbake til steg 2 når brukeren klikker på Tilbake', async () => {
-      const user = userEvent.setup()
-      const navigateMock = vi.fn()
-      vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
-        () => navigateMock
-      )
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      expect(
-        await screen.findByText('stegvisning.offentligtp.title')
-      ).toBeVisible()
-
-      await user.click(await screen.findByText('stegvisning.tilbake'))
-
-      expect(navigateMock).toHaveBeenCalledWith(paths.samtykke)
-    })
-
-    it('nullstiller input fra brukeren og redirigerer til landingssiden når brukeren klikker på Avbryt', async () => {
-      const flushMock = vi.spyOn(
-        userInputReducerUtils.userInputActions,
-        'flush'
-      )
-      const user = userEvent.setup()
-      const navigateMock = vi.fn()
-      vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
-        () => navigateMock
-      )
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      expect(
-        await screen.findByText('stegvisning.offentligtp.title')
-      ).toBeVisible()
-
-      await user.click(await screen.findByText('stegvisning.avbryt'))
-
-      expect(navigateMock).toHaveBeenCalledWith(paths.login)
-      expect(flushMock).toHaveBeenCalled()
+    await waitFor(async () => {
+      expect(await screen.findByTestId('step3-loader')).toBeVisible()
+      expect(document.title).toBe('application.title.stegvisning.step3')
     })
   })
 
-  describe('Gitt at kallet til /tpo-medlemskap feiler', async () => {
-    it('rendrer tekstene riktig', async () => {
-      mockErrorResponse('/tpo-medlemskap')
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      expect(
-        await screen.findByText('stegvisning.offentligtp.error.title')
-      ).toBeVisible()
-      expect(
-        await screen.findByText(
-          'Gå videre for å beregne alderspensjon fra NAV',
-          { exact: false }
-        )
-      ).toBeVisible()
-      expect(await screen.findByText('stegvisning.neste')).toBeVisible()
-      expect(await screen.findByText('stegvisning.tilbake')).toBeVisible()
-      expect(await screen.findByText('stegvisning.avbryt')).toBeVisible()
+  it('rendrer Step 3 slik den skal', async () => {
+    const router = createMemoryRouter(routes, {
+      basename: BASE_PATH,
+      initialEntries: [`${BASE_PATH}${paths.afp}`],
+    })
+    render(<RouterProvider router={router} />, {
+      hasRouter: false,
+    })
+    expect(await screen.findByRole('heading', { level: 2 })).toHaveTextContent(
+      'stegvisning.afp.title'
+    )
+    expect(screen.getAllByRole('radio')).toHaveLength(4)
+  })
+
+  it('Når brukeren velger afp og klikker på Neste, registrerer afp og navigerer videre til step 4', async () => {
+    const setAfpMock = vi.spyOn(
+      userInputReducerUtils.userInputActions,
+      'setAfp'
+    )
+    const user = userEvent.setup()
+
+    const navigateMock = vi.fn()
+    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
+      () => navigateMock
+    )
+    const router = createMemoryRouter(routes, {
+      basename: BASE_PATH,
+      initialEntries: [`${BASE_PATH}${paths.afp}`],
+    })
+    render(<RouterProvider router={router} />, {
+      hasRouter: false,
     })
 
-    it('sender videre til steg 4 når brukeren klikker på Neste', async () => {
-      mockErrorResponse('/tpo-medlemskap')
-      const user = userEvent.setup()
-      const navigateMock = vi.fn()
-      vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
-        () => navigateMock
-      )
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      expect(
-        screen.queryByText('stegvisning.offentligtp.title')
-      ).not.toBeInTheDocument()
-      const button = await screen.findByText('stegvisning.neste')
+    const radioButtons = await screen.findAllByRole('radio')
+    await user.click(radioButtons[0])
 
-      await user.click(button)
+    await user.click(screen.getByText('stegvisning.neste'))
 
-      expect(navigateMock).toHaveBeenCalledWith(paths.afp)
+    expect(setAfpMock).toHaveBeenCalledWith('ja_offentlig')
+    expect(navigateMock).toHaveBeenCalledWith(paths.ufoeretrygdAFP)
+  })
+
+  it('sender tilbake til steg 2 når brukeren klikker på Tilbake', async () => {
+    const user = userEvent.setup()
+    const navigateMock = vi.fn()
+    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
+      () => navigateMock
+    )
+    const router = createMemoryRouter(routes, {
+      basename: BASE_PATH,
+      initialEntries: [`${BASE_PATH}${paths.afp}`],
+    })
+    render(<RouterProvider router={router} />, {
+      hasRouter: false,
+    })
+    await waitFor(async () => {
+      await user.click(await screen.findByText('stegvisning.tilbake'))
+    })
+    expect(navigateMock).toHaveBeenCalledWith(paths.utenlandsopphold)
+  })
+
+  it('nullstiller input fra brukeren og redirigerer til landingssiden når brukeren klikker på Avbryt', async () => {
+    const flushMock = vi.spyOn(userInputReducerUtils.userInputActions, 'flush')
+    const user = userEvent.setup()
+
+    const navigateMock = vi.fn()
+    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
+      () => navigateMock
+    )
+    const router = createMemoryRouter(routes, {
+      basename: BASE_PATH,
+      initialEntries: [`${BASE_PATH}${paths.afp}`],
+    })
+    render(<RouterProvider router={router} />, {
+      hasRouter: false,
     })
 
-    it('sender tilbake til steg 2 når brukeren klikker på Tilbake', async () => {
-      mockErrorResponse('/tpo-medlemskap')
-      const user = userEvent.setup()
-      const navigateMock = vi.fn()
-      vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
-        () => navigateMock
-      )
-      const router = createMemoryRouter(routes, {
-        basename: BASE_PATH,
-        initialEntries: [`${BASE_PATH}${paths.offentligTp}`],
-      })
-      render(<RouterProvider router={router} />, {
-        hasRouter: false,
-      })
-      expect(
-        screen.queryByText('stegvisning.offentligtp.title')
-      ).not.toBeInTheDocument()
-      const button = await screen.findByText('stegvisning.tilbake')
+    const radioButtons = await screen.findAllByRole('radio')
+    await user.click(radioButtons[0])
 
-      await user.click(button)
+    await user.click(screen.getByText('stegvisning.avbryt'))
 
-      expect(navigateMock).toHaveBeenCalledWith(paths.samtykke)
-    })
+    expect(navigateMock).toHaveBeenCalledWith(paths.login)
+    expect(flushMock).toHaveBeenCalled()
   })
 })
