+++ src/pages/Step4/Step4.tsx
@@ -1,30 +1,17 @@
 import React from 'react'
 import { useIntl } from 'react-intl'
-import { useNavigate, Await } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
-import { Loader } from '@/components/common/Loader'
-import { AFP } from '@/components/stegvisning/AFP'
+import { Ufoere } from '@/components/stegvisning/Ufoere'
 import { paths } from '@/router/constants'
-import { useStep4AccessData } from '@/router/loaders'
-import { useGetTpoMedlemskapQuery } from '@/state/api/apiSlice'
 import { useAppDispatch, useAppSelector } from '@/state/hooks'
-import {
-  selectSamtykke,
-  selectAfp,
-  selectIsVeileder,
-} from '@/state/userInput/selectors'
+import { selectIsVeileder } from '@/state/userInput/selectors'
 import { userInputActions } from '@/state/userInput/userInputReducer'
 
 export function Step4() {
   const intl = useIntl()
   const navigate = useNavigate()
   const dispatch = useAppDispatch()
-  const loaderData = useStep4AccessData()
-  const harSamtykket = useAppSelector(selectSamtykke)
-  const previousAfp = useAppSelector(selectAfp)
-
-  const { data: TpoMedlemskap, isSuccess: isTpoMedlemskapQuerySuccess } =
-    useGetTpoMedlemskapQuery(undefined, { skip: !harSamtykket })
 
   const isVeileder = useAppSelector(selectIsVeileder)
 
@@ -43,47 +30,12 @@ export function Step4() {
       }
 
   const onPrevious = (): void => {
-    if (
-      isTpoMedlemskapQuerySuccess &&
-      TpoMedlemskap.harTjenestepensjonsforhold
-    ) {
-      return navigate(paths.offentligTp)
-    } else {
-      return navigate(paths.samtykke)
-    }
+    return navigate(paths.afp)
   }
 
-  const onNext = (afpData: AfpRadio): void => {
-    dispatch(userInputActions.setAfp(afpData))
-    navigate(paths.ufoeretrygdAFP)
+  const onNext = (): void => {
+    navigate(paths.sivilstand)
   }
 
-  return (
-    <React.Suspense
-      fallback={
-        <div style={{ width: '100%' }}>
-          <Loader
-            data-testid="step4-loader"
-            size="3xlarge"
-            title={intl.formatMessage({ id: 'pageframework.loading' })}
-            isCentered
-          />
-        </div>
-      }
-    >
-      <Await resolve={loaderData.shouldRedirectTo}>
-        {(shouldRedirectTo: string) => {
-          return (
-            <AFP
-              shouldRedirectTo={shouldRedirectTo}
-              afp={previousAfp}
-              onCancel={onCancel}
-              onPrevious={onPrevious}
-              onNext={onNext}
-            />
-          )
-        }}
-      </Await>
-    </React.Suspense>
-  )
+  return <Ufoere onCancel={onCancel} onPrevious={onPrevious} onNext={onNext} />
 }
