+++ src/pages/Step6/Step6.tsx
@@ -2,11 +2,14 @@ import React from 'react'
 import { useIntl } from 'react-intl'
 import { useNavigate } from 'react-router-dom'
 
-import { SamtykkeOffentligAFP } from '@/components/stegvisning/SamtykkeOffentligAFP'
+import { Samtykke } from '@/components/stegvisning/Samtykke'
 import { paths } from '@/router/constants'
+import { apiSlice, useGetUfoeregradQuery } from '@/state/api/apiSlice'
 import { useAppDispatch, useAppSelector } from '@/state/hooks'
 import {
-  selectSamtykkeOffentligAFP,
+  selectAfp,
+  selectSamtykke,
+  selectHarHentetTpoMedlemskap,
   selectIsVeileder,
 } from '@/state/userInput/selectors'
 import { userInputActions } from '@/state/userInput/userInputReducer'
@@ -15,16 +18,27 @@ export function Step6() {
   const intl = useIntl()
   const navigate = useNavigate()
   const dispatch = useAppDispatch()
-  const harSamtykketOffentligAFP = useAppSelector(selectSamtykkeOffentligAFP)
+  const harSamtykket = useAppSelector(selectSamtykke)
+  const afp = useAppSelector(selectAfp)
+  const shouldFlush = useAppSelector(selectHarHentetTpoMedlemskap)
   const isVeileder = useAppSelector(selectIsVeileder)
 
+  const { data: ufoeregrad } = useGetUfoeregradQuery()
   React.useEffect(() => {
     document.title = intl.formatMessage({
       id: 'application.title.stegvisning.step6',
     })
   }, [])
 
-  // Fjern mulighet for avbryt hvis person er veileder
+  const onNext = (samtykkeData: BooleanRadio) => {
+    const samtykke = samtykkeData === 'ja'
+    dispatch(userInputActions.setSamtykke(samtykke))
+    if (shouldFlush && !samtykke) {
+      apiSlice.util.invalidateTags(['TpoMedlemskap', 'Pensjonsavtaler'])
+    }
+    navigate(paths.beregningEnkel)
+  }
+
   const onCancel = isVeileder
     ? undefined
     : (): void => {
@@ -33,18 +47,18 @@ export function Step6() {
       }
 
   const onPrevious = (): void => {
-    return navigate(paths.afp)
-  }
-
-  const onNext = (samtykkeData: BooleanRadio) => {
-    const samtykke = samtykkeData === 'ja'
-    dispatch(userInputActions.setSamtykkeOffentligAFP(samtykke))
-    navigate(paths.sivilstand)
+    if (ufoeregrad?.ufoeregrad && afp && afp !== 'nei') {
+      navigate(paths.ufoeretrygdAFP)
+    } else if (ufoeregrad?.ufoeregrad === 0 && afp === 'ja_offentlig') {
+      navigate(paths.samtykkeOffentligAFP)
+    } else {
+      navigate(paths.afp)
+    }
   }
 
   return (
-    <SamtykkeOffentligAFP
-      harSamtykket={harSamtykketOffentligAFP}
+    <Samtykke
+      harSamtykket={harSamtykket}
       onCancel={onCancel}
       onPrevious={onPrevious}
       onNext={onNext}
