+++ src/pages/Step6/__tests__/Step6.test.tsx
@@ -3,71 +3,194 @@ import * as ReactRouterUtils from 'react-router'
 import { describe, it, vi } from 'vitest'
 
 import { Step6 } from '..'
+import {
+  fulfilledGetTpoMedlemskap,
+  fulfilledGetUfoeregrad,
+  fulfilledPensjonsavtaler,
+} from '@/mocks/mockedRTKQueryApiCalls'
+import { mockResponse } from '@/mocks/server'
 import { paths } from '@/router/constants'
+import { apiSlice } from '@/state/api/apiSlice'
+import * as apiSliceUtils from '@/state/api/apiSlice'
+import { store as stateStore } from '@/state/store'
+import { selectHarHentetTpoMedlemskap } from '@/state/userInput/selectors'
 import { userInputInitialState } from '@/state/userInput/userInputReducer'
-import { screen, render, userEvent } from '@/test-utils'
+import { screen, render, userEvent, waitFor } from '@/test-utils'
 
 describe('Step 6', () => {
-  it('har riktig sidetittel', () => {
+  it('har riktig sidetittel', async () => {
     render(<Step6 />)
     expect(document.title).toBe('application.title.stegvisning.step6')
   })
 
-  describe('Gitt at brukeren svarer Ja på spørsmål om samtykke til beregning av offentlig AFP', async () => {
+  describe('Gitt at brukeren svarer Ja på spørsmål om samtykke', async () => {
     it('registrerer samtykke og navigerer videre til riktig side når brukeren klikker på Neste', async () => {
       const user = userEvent.setup()
       const navigateMock = vi.fn()
       vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
         () => navigateMock
       )
-      const { store } = render(<Step6 />, {})
+
+      const { store } = render(<Step6 />)
       const radioButtons = screen.getAllByRole('radio')
 
       await user.click(radioButtons[0])
-      await user.click(screen.getByText('stegvisning.neste'))
+      await user.click(screen.getByText('stegvisning.beregn'))
 
-      expect(store.getState().userInput.samtykkeOffentligAFP).toBe(true)
-      expect(navigateMock).toHaveBeenCalledWith(paths.sivilstand)
+      expect(store.getState().userInput.samtykke).toBe(true)
+      expect(navigateMock).toHaveBeenCalledWith(paths.beregningEnkel)
     })
   })
 
-  describe('Gitt at brukeren svarer Nei på spørsmål om samtykke til beregning av offentlig AFP', async () => {
-    it('registrerer samtykke og navigerer videre til riktig side når brukeren klikker på Neste', async () => {
+  describe('Gitt at brukeren svarer Nei på spørsmål om samtykke', async () => {
+    it('invaliderer cache for tpo-medlemskap og pensjonsavtaler i storen (for å fjerne evt. data som ble hentet pga en tidligere samtykke). Navigerer videre til riktig side når brukeren klikker på Neste', async () => {
+      const invalidateMock = vi.spyOn(
+        apiSliceUtils.apiSlice.util.invalidateTags,
+        'match'
+      )
+
       const user = userEvent.setup()
       const navigateMock = vi.fn()
       vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
         () => navigateMock
       )
-      const { store } = render(<Step6 />, {})
+
+      const { store } = render(<Step6 />, {
+        preloadedState: {
+          api: {
+            /* eslint-disable @typescript-eslint/ban-ts-comment */
+            // @ts-ignore
+            queries: {
+              ...fulfilledGetTpoMedlemskap,
+              ...fulfilledPensjonsavtaler,
+              ...fulfilledGetUfoeregrad,
+            },
+          },
+          userInput: {
+            ...userInputInitialState,
+            samtykke: true,
+          },
+        },
+      })
+      await store.dispatch(
+        apiSliceUtils.apiSlice.endpoints.getTpoMedlemskap.initiate()
+      )
+      expect(Object.keys(store.getState().api.queries).length).toEqual(3)
+
+      expect(selectHarHentetTpoMedlemskap(store.getState())).toBe(true)
+
       const radioButtons = screen.getAllByRole('radio')
 
       await user.click(radioButtons[1])
-      await user.click(screen.getByText('stegvisning.neste'))
+      await user.click(screen.getByText('stegvisning.beregn'))
 
-      expect(store.getState().userInput.samtykkeOffentligAFP).toBe(false)
-      expect(navigateMock).toHaveBeenCalledWith(paths.sivilstand)
+      expect(store.getState().userInput.samtykke).toBe(false)
+      expect(invalidateMock).toHaveBeenCalledTimes(2)
+
+      expect(navigateMock).toHaveBeenCalledWith(paths.beregningEnkel)
     })
   })
 
-  it('nullstiller input fra brukeren og sender tilbake til steg 4 når brukeren klikker på Tilbake', async () => {
+  it('sender tilbake til steg 5 når brukeren som mottar uføretrygd og som har valgt afp, klikker på Tilbake', async () => {
     const user = userEvent.setup()
+    mockResponse('/v1/ufoeregrad', {
+      status: 200,
+      json: {
+        ufoeregrad: 100,
+      },
+    })
     const navigateMock = vi.fn()
     vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
       () => navigateMock
     )
-    const { store } = render(<Step6 />, {
+
+    render(<Step6 />, {
       preloadedState: {
-        userInput: { ...userInputInitialState, samtykkeOffentligAFP: null },
+        userInput: { ...userInputInitialState, afp: 'ja_offentlig' },
+      },
+      hasRouter: false,
+    })
+
+    await stateStore.dispatch(apiSlice.endpoints.getUfoeregrad.initiate())
+
+    await waitFor(async () => {
+      await user.click(screen.getByText('stegvisning.tilbake'))
+      expect(navigateMock).toHaveBeenCalledWith(paths.ufoeretrygdAFP)
+    })
+  })
+
+  it('sender tilbake til steg 4 når brukeren som mottar uføretrygd og har ikke valgt afp, klikker på Tilbake', async () => {
+    const user = userEvent.setup()
+    mockResponse('/v1/ufoeregrad', {
+      status: 200,
+      json: {
+        ufoeregrad: 100,
       },
     })
-    const radioButtons = screen.getAllByRole('radio')
+    const navigateMock = vi.fn()
+    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
+      () => navigateMock
+    )
 
-    await user.click(radioButtons[0])
-    expect(radioButtons[0]).toBeChecked()
-    await user.click(screen.getByText('stegvisning.tilbake'))
+    render(<Step6 />, {
+      preloadedState: {
+        userInput: { ...userInputInitialState, afp: 'nei' },
+      },
+      hasRouter: false,
+    })
 
-    expect(navigateMock).toHaveBeenCalledWith(paths.afp)
-    expect(store.getState().userInput.samtykkeOffentligAFP).toBe(null)
+    await stateStore.dispatch(apiSlice.endpoints.getUfoeregrad.initiate())
+
+    await waitFor(async () => {
+      await user.click(screen.getByText('stegvisning.tilbake'))
+      expect(navigateMock).toHaveBeenCalledWith(paths.afp)
+    })
+  })
+
+  it('sender tilbake til steg 6 når brukeren som ikke mottar uføretrygd og som har valgt AFP offentlig klikker på Tilbake', async () => {
+    const user = userEvent.setup()
+
+    const navigateMock = vi.fn()
+    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
+      () => navigateMock
+    )
+
+    render(<Step6 />, {
+      preloadedState: {
+        userInput: { ...userInputInitialState, afp: 'ja_offentlig' },
+      },
+      hasRouter: false,
+    })
+
+    await stateStore.dispatch(apiSlice.endpoints.getUfoeregrad.initiate())
+
+    await waitFor(async () => {
+      await user.click(screen.getByText('stegvisning.tilbake'))
+      expect(navigateMock).toHaveBeenCalledWith(paths.samtykkeOffentligAFP)
+    })
+  })
+
+  it('sender tilbake til steg 4 når brukeren som ikke mottar uføretrygd og som har valgt noe annet enn har valgt AFP offentlig, klikker på Tilbake', async () => {
+    const user = userEvent.setup()
+
+    const navigateMock = vi.fn()
+    vi.spyOn(ReactRouterUtils, 'useNavigate').mockImplementation(
+      () => navigateMock
+    )
+
+    render(<Step6 />, {
+      preloadedState: {
+        userInput: { ...userInputInitialState, afp: 'ja_privat' },
+      },
+      hasRouter: false,
+    })
+
+    await stateStore.dispatch(apiSlice.endpoints.getUfoeregrad.initiate())
+
+    await waitFor(async () => {
+      await user.click(screen.getByText('stegvisning.tilbake'))
+      expect(navigateMock).toHaveBeenCalledWith(paths.afp)
+    })
   })
 
   it('nullstiller input fra brukeren og redirigerer til landingssiden når brukeren klikker på Avbryt', async () => {
@@ -78,7 +201,7 @@ describe('Step 6', () => {
     )
     const { store } = render(<Step6 />, {
       preloadedState: {
-        userInput: { ...userInputInitialState, samtykkeOffentligAFP: false },
+        userInput: { ...userInputInitialState, samtykke: false },
       },
     })
     const radioButtons = screen.getAllByRole('radio')
@@ -87,6 +210,6 @@ describe('Step 6', () => {
     await user.click(screen.getByText('stegvisning.avbryt'))
 
     expect(navigateMock).toHaveBeenCalledWith(paths.login)
-    expect(store.getState().userInput.samtykkeOffentligAFP).toBe(null)
+    expect(store.getState().userInput.samtykke).toBe(null)
   })
 })
