--- src/pages/Step7/index.ts
@@ -1,1 +1,0 @@
-export { Step7 } from './Step7'
