+++ src/translations/en.ts
@@ -7,7 +7,6 @@ const translations = {
   'application.title.stegvisning.step4': 'MANGLER_TEKST',
   'application.title.stegvisning.step5': 'MANGLER_TEKST',
   'application.title.stegvisning.step6': 'MANGLER_TEKST',
-  'application.title.stegvisning.step7': 'MANGLER_TEKST',
   'application.title.stegvisning.uventet_feil': 'MANGLER_TEKST',
   'application.title.beregning': 'MANGLER_TEKST',
   'application.title.forbehold': 'MANGLER_TEKST',
@@ -75,11 +74,6 @@ const translations = {
   'stegvisning.samtykke.radio_ja': 'MANGLER_TEKST',
   'stegvisning.samtykke.radio_nei': 'MANGLER_TEKST',
   'stegvisning.samtykke.validation_error': 'MANGLER_TEKST',
-  'stegvisning.offentligtp.title': 'MANGLER_TEKST',
-  'stegvisning.offentligtp.loading': 'MANGLER_TEKST',
-  'stegvisning.offentligtp.ingress': 'MANGLER_TEKST',
-  'stegvisning.offentligtp.error.title': 'MANGLER_TEKST',
-  'stegvisning.offentligtp.error.ingress': 'MANGLER_TEKST',
   'stegvisning.afp.title': 'MANGLER_TEKST',
   'stegvisning.afp.ingress': 'MANGLER_TEKST',
   'stegvisning.afp.readmore_privat_title': 'MANGLER_TEKST',
